handle abs paths as argv[0]

+++ pidfile.py
@@ -6,10 +6,17 @@ import atexit
 
 pidfile = None
 
-def write():
+def write(reset = False):
     global pidfile
     pid = str(os.getpid())
-    pidfile = os.path.join("/tmp/", os.path.splitext(sys.argv[0])[0] + ".pid")
+
+    pname = os.path.splitext(os.path.split(sys.argv[0])[-1])[0]
+    pidfile = os.path.join("/tmp/", pname + ".pid")
+
+    if reset == False and os.path.exists(pidfile):
+        logging.error("pid file exists, kill running procs and remove {}".format(pidfile))
+        exit(-1)
+
     logging.info('wrote pid to {}'.format(pidfile))
     with open(pidfile,'w')as f:
         f.write(pid)
@@ -17,6 +24,7 @@ def write():
     atexit.register(cleanup)
 def cleanup():
     global pidfile
+    logging.info('cleanup pidfile: {}'.format(pidfile))
     if pidfile != None:
         os.unlink(pidfile)
         pidfile = None
